Feat 05:
One new functionality is added to the Buy Books menu.

menu number 3 is a new method that shows the list of the book item in
the basket.

[ 3 ]  Show the basket items

Since the newly-added function is  added as item number 3, the previous
menu item number 3 moved to item number 4.

All Buy Books menu is:
[ 1 ]  Add a book to your basket
[ 2 ]  Remove a book from your basket
[ 3 ]  Show the basket items          
[ 4 ]  Buy the current basket

( 0 )  Exit (Drop the current order)